+++ packages/can/can.iterate/src/lib/canIterate.ts
@@ -1,6 +1,6 @@
 import { isNil } from '@zerodep/is.nil';
 
-export function isIterable(value: any): boolean {
+export function canIterate(value: any): boolean {
   // we do not want any kind of number or string to be iterable
   const type = typeof value;
   if (type === 'string' || isNil(value)) {