+++ packages/can/can/src/index.spec.ts
@@ -1,0 +1,8 @@
+import * as utils from './index';
+
+describe('Utils barrel package', () => {
+  it('should export specific packages and interfaces', () => {
+    // console.log(Object.keys(utils))
+    expect(Object.keys(utils)).toEqual(['canIterate']);
+  });
+});