+++ packages/is/is.error/src/lib/isError.spec.ts
@@ -10,13 +10,13 @@ const positiveCases = [
 
 describe('isInteger', () => {
   // @ts-ignore
-  test.each(positiveCases)('should allow a/an %s', (title, value) => {
+  test.each(positiveCases)('should return TRUE for a/an %s', (title, value) => {
     // @ts-ignore
     expect(isError(value)).toEqual(true);
   });
 
   // @ts-ignore
-  test.each(negativeCases)('should NOT allow a/an %s', (title, value) => {
+  test.each(negativeCases)('should return FALSE for a/an %s', (title, value) => {
     // @ts-ignore
     expect(isError(value)).toEqual(false);
   });