+++ packages/utils/src/index.ts
@@ -1,3 +1,4 @@
+export * from '@zerodep/can';
 export * from '@zerodep/errors';
 export * from '@zerodep/guard';
 export * from '@zerodep/is';